Mask YAML files
