--- scripts/tests/samples/secret1.yaml
@@ -1,44 +1,0 @@
-parameter_defaults:
-  AdminPassword: dummyvalue
-  AdminToken: dummyvalue
-  AodhPassword: dummyvalue
-  HeatStackDomainAdminPassword: dummyvalue
-  HorizonSecret: dummyvalue
-  IronicPassword: dummyvalue
-  KeystoneCredential0: testpass
-  KeystoneCredential1: testpass
-  KeystoneFernetKey0: testpass
-  KeystoneFernetKey1: testpass
-  MigrationSshKey:
-    private_key: '-----BEGIN RSA PRIVATE KEY-----
-
-      dummy value
-
-      -----END RSA PRIVATE KEY-----
-
-      '
-    public_key: ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQCj7vO8eBgY5uoU4SheQENoon6NegupSYeQ5OZwFYM9alTMXXapk4Qq1aZbkEsKfNwqL8ZdURUAxHO94SktFf8/OwoI0aNQur7kHWP2x8fEvhE6FtC5xKsOU+jasn8zooTnAhFEv9MLG5HTzPhSZkdMcAtrMjtas3e1kBWhfkVJVIUrQbjelksf1E7l1wADCYxErcwpuSsgaKxv/3M2kDoW1TF4Z1Deb7eY4q87rgpcCMWQ4PihWAhfHpGSZ+GbsSA0KOG1agQIvsqUidFwiaJWsGAyB+WooYa/znhPYNdVQcLoNa7ajHDeWqB6aRCYMVRkbieoMmhNRvJsfpe7JxVx
-      Generated by TripleO
-  MysqlClustercheckPassword: dummyvalue
-  MysqlMariabackupPassword: dummyvalue
-  MysqlRootPassword: dummyvalue
-  NeutronMetadataProxySharedSecret: dummyvalue
-  NeutronPassword: dummyvalue
-  NotifyPassword: dummyvalue
-  NovaPassword: dummyvalue
-  NovajoinPassword: dummyvalue
-  OctaviaCaKeyPassphrase: dummyvalue
-  OctaviaHeartbeatKey: dummyvalue
-  OctaviaPassword: dummyvalue
-  OctaviaServerCertsKeyPassphrase: dummyvalue
-  PankoPassword: dummyvalue
-  PcsdPassword: dummyvalue
-  PlacementPassword: dummyvalue
-  RabbitCookie: dummyvalue
-  RabbitPassword: dummyvalue
-  RedisPassword: dummyvalue
-  RpcPassword: dummyvalue
-  SaharaPassword: dummyvalue
-  SnmpdReadonlyUserPassword: dummyvalue
-  SwiftHashSuffix: dummyvalue
-  SwiftPassword: dummyvalue
