--- scripts/tests/samples/secret2.yaml
@@ -1,44 +1,0 @@
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-resources:
-labels:
-  - pairs:
-      created-by: install_yamls
-secretGenerator:
-- name: osp-secret
-  literals:
-  - AdminPassword=12345678
-  - AodhPassword=12345678
-  - BarbicanPassword=12345678
-  - BarbicanSimpleCryptoKEK=mypass
-  - CeilometerPassword=12345678
-  - DbRootPassword=12345678
-  - DatabasePassword=12345678
-  - DesignatePassword=12345678
-  - PlacementPassword=12345678
-  - GlancePassword=12345678
-  - NeutronPassword=12345678
-  - CinderPassword=12345678
-  - IronicPassword=12345678
-  - IronicInspectorPassword=12345678
-  - KeystoneClientSecret=mysecret
-  - KeystoneCryptoPassphrase=dummyvalue
-  - OctaviaPassword=12345678
-  - OctaviaHeartbeatKey=12345678
-  - NovaPassword=12345678
-  - ManilaPassword=12345678
-  - MetadataSecret=1234567842
-  - HeatPassword=12345678
-  - HeatAuthEncryptionKey=mykey
-  - HeatStackDomainAdminPassword=12345678
-  - SwiftPassword=12345678
-- name: libvirt-secret
-  literals:
-  - LibvirtPassword=12345678
-- name: octavia-ca-passphrase
-  literals:
-  - server-ca-passphrase=12345678
-generatorOptions:
-  disableNameSuffixHash: true
-  labels:
-    type: osp-secret
